fix(mapper): treat zeroPaddedAddress() as a transformation
zeroPaddedAddress() was being treated as an abbreviator, but there is
no guesswork involved as it left-pads a full address.  So split this
out into a new transformers.ts file, and substitute zero-padded
addresses with the normal format not the guess format.

Also substitute lowercase zero-padded addresses, as seen on the likes
of polygonscan in the Events tab of smart contract pages.
