+++ src/shared/abbreviators.ts
@@ -22,13 +22,4 @@ export function abbreviatedAddresses(address: string): string[] {
   );
 }
 
-// Spot addresses embedded into EVM 256 bit words.  This is typically
-// seen in event data.
-export function zeroPaddedAddress(address: string): string[] {
-  // Addresses are 20 bytes, so to pad to 32 bytes we need an extra
-  // 12 bytes of zero padding (24 nibbles).
-  const padded = '00'.repeat(12) + address.slice(2);
-  return [padded, '0x' + padded];
-}
-
-export const ABBREVIATION_FUNCTIONS = [abbreviatedAddresses, zeroPaddedAddress];
+export const ABBREVIATION_FUNCTIONS = [abbreviatedAddresses];
