+++ src/shared/transformers.ts
@@ -1,0 +1,11 @@
+// Spot addresses embedded into EVM 256 bit words.  This is typically
+// seen in event data.
+export function zeroPaddedAddress(address: string): string[] {
+  // Addresses are 20 bytes, so to pad to 32 bytes we need an extra
+  // 12 bytes of zero padding (24 nibbles).
+  const padded = '00'.repeat(12) + address.slice(2);
+  const addrs = [padded, '0x' + padded];
+  return [...addrs, ...addrs.map(a => a.toLowerCase())];
+}
+
+export const TRANSFORMER_FUNCTIONS = [zeroPaddedAddress];
